Ensure proper casting for int16

+++ Matlab/rshift.m
@@ -1,5 +1,5 @@
 % Perform bitwise right shift of input a on k bits 
 % Equivalent to a>>k 
 function out = rshift(a,k) 
-out = a / 2^k;
+out = double(int16(a) / 2^k);
 end
